fixing syntax to make python3 to work

+++ src/utils/utils.py
@@ -114,9 +114,8 @@ def tps_warp(Y, T, Y_image, out_shape):
 def checkboard(I1, I2, n=7):
     assert I1.shape == I2.shape
     height, width, channels = I1.shape
-    hi, wi = height/n, width/n
-    outshape = (hi*n, wi*n, channels)
-
+    hi, wi = int(height/n), int(width/n)
+    outshape = (hi*n, wi*n, int(channels))
     out_image = np.zeros(outshape, dtype='uint8')
     for i in range(n):
         h = hi * i
